requirements.txt
asgiref==3.7.2
Django==5.0.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
phonenumberslite==8.13.28
pytz==2023.3.post1
sqlparse==0.4.4
